+++ CHANGELOG.md
@@ -30,6 +30,7 @@
  - fix translation extractor code and update translations
  - add shadow to copy button #491
  - fix empty state messages in photos collection by theme #466
+ - send mail to user when subscription to event is declined #377
 
 ## 0.1.0 (2017-12-13)
  - production deployment